# Optimize and refine code to c++17