--- a/.gitignore
+++ b/.gitignore
@@ -2,6 +2,7 @@ CMakeCache.txt
 CMakeFiles/
 Makefile
 *.cmake
+.vscode/
 
 emphf_config.hpp
 compute_mphf_hem
@@ -11,4 +12,5 @@ compute_mphf_seq
 gen_synthetic_data
 test_mphf
 test_mphf_hem
-
+test.dat
+test.pf