--- a/compute_mphf_seq.cpp
+++ b/compute_mphf_seq.cpp
@@ -2,10 +2,12 @@
 #include "hypergraph.hpp"
 #include "hypergraph_sorter_seq.hpp"
 
-int main(int argc, char** argv)
+int main(int argc, char** argv) noexcept
 {
     using namespace emphf;
-    return compute_mphf_main<hypergraph_sorter_seq<hypergraph<uint32_t>>,
-                             hypergraph_sorter_seq<hypergraph<uint64_t>>,
-                             jenkins64_hasher>(argc, argv);
+    return compute_mphf_main<
+        hypergraph_sorter_seq<hypergraph<uint32_t>>,
+        hypergraph_sorter_seq<hypergraph<uint64_t>>,
+        jenkins64_hasher
+    >(argc, argv);
 }